Add a LICENSE.txt file.
Opting for the modified BSD license for simplicity. This project is just about as open as I can stand it. If it helps anyone learn more about the z80 systems, or if someone decides to port it to some esoteric platform I've never heard of, I'd like to hear about it!