Hook up audio sample generation for Channel 2. (LOUD!)
This simply generates a frequency oscillator for Channel 2's data,
but it sets the framework for the rest of the system. I'm limited
a bit by Love2D here, as it doesn't have a proper circular audio
buffer, so I'll never be able to fully get rid of clicks and pops.
Still, it sounds like the frequency data is accurate, so that's
good!