Color: implement VRAM banks and bank switching, some color games boot
This needs a *lot* of work, and there's a ton of graphical glitches,
but I'm still impressed that so many games magically started booting
with just these two features, and a bit of the detection logic.

Of course, this causes regressions on any gameboy color games that
support B&W mode, as it's defaulting to a broken color implementation.