First steps to dynamic linking.
Audio is now a proper module, and could theoretically be swapped out
with an alternate implementation. Other modules will follow suit,
this commit lays the groundwork for the rewrite.

Converts the gameboy module into a proper Class, though a lot of work
still needs to be done to properly remove "module scope" programming
and package everything up nicely.