Remove gameboy reference from filebrowser. Fixes palette selection.
Basically the filebrowser's gameboy reference was getting stale,
causing it to apply the palette changes to an old copy, rather than
the active game. Instead, we keep the palette selection within the
filebrowser, and let LuaGB stitch the values together. This decouples
the filebrowser from the gameboy module entirely, which is ideal.