Correct hblank and general purpose DMA to be much more accurate
I was following incomplete documentation, and wasn't aware that
hblank DMA can be triggered right away if started during hblank
or while the display is disabled. Some games seem to utilize this
quirk for very fast loading of VRAM tiles. Other games rely on the
destination and source registers retaining their old values, and
the startup calculation for this was incorrect, leading to other
bugs.

This culminates in Dragon's Lair booting for the first time. It
now appears to be playable, though I'm not very good at it, so
I couldn't get very far into it to test.