Make the parsed header available to MBC controllers
This is mostly to facilitate easier filtering of selected banks, in
case games try to switch to a bank that doesn't exist, so that
mirroring happens correctly. Right now only MBC5 is making use
of this feature, to disable its 9th bank selection bit on <4MB carts.
This didn't end up fixing Dragon's Lair, but it's worth keeping anyway
since the MBC code is a little too trusting in its current form.