Several register writes only have an effect on a note trigger
Note: I determined this behavior experimentally. I'm committing it
to not lose the work (since the behavior is now correct) but it
does invalidate a good chunk of register write tests, which need
to be updated. Specifically, the new business logic is:

Some register writes have NO effect when written to io.ram
Writes to NRx4 with the "trigger" bit set read new values from io.ram

This makes sense intuitively, as partial updates shouldn't be able
to cause the previously playing note to glitch audibly due to the
time delay between register writes.