Writes to waveform duty should NOT affect position
The waveform has a base form (represented as a byte), and then a
position into that waveform that the period timer controls.
I incorrectly assumed that writes to the waveform would simply
replace the byte; this is incorrect, and doing so causes games
that update the audio registers every frame to introduce a 60 Hz
drone as the waveform is reset mid-period. This is quite audible
and sounds horrifying.

This experimental fix alters the waveform and position logic to
account for this. It breaks some tests, which expect the waveform
to shift in place; they should be rewritten to verify the
position's behavior and its effect on the output.