Delay vBlank interrupt by at least a handful of cycles
This implementation delays it by just about 1 instruction, which is
enough to work around certain games with CPU spinwaits on LY=144.
If such a game has vBlank interrupts enabled, by documentation the
vBlank interrupt takes priority... but in practice there's like
a 2-4 cycle window where their spin-wait will see LY=144 before
the vBlank interrupt fires.

The only "game" I'm aware of that relies on this behavior is
the excellent Pocket demo:
http://www.pouet.net/prod.php?which=65997

That demo now makes it all the way through, and most of the effects
even work! There's only one glitchy section remaining and that's
going to have to wait until the GPU rewrite...